+++ packages/css-color-parser/src/index.ts
@@ -15,6 +15,7 @@ import { oklab } from './functions/oklab';
 import { oklch } from './functions/oklch';
 import { rgb } from './functions/rgb';
 import { toLowerCaseAZ } from './util/to-lower-case-a-z';
+import { contrastColor } from './functions/contrast-color';
 
 export type { ColorData } from './color-data';
 export { ColorNotation } from './color-notation';
@@ -49,6 +50,8 @@ export function color(colorNode: ComponentValue): ColorData | false {
 				return colorFn(colorNode, color);
 			case 'color-mix':
 				return colorMix(colorNode, color);
+			case 'contrast-color':
+				return contrastColor(colorNode, color);
 		}
 	}