+++ plugins/postcss-color-function/src/index.ts
@@ -42,6 +42,10 @@ const basePlugin: PluginCreator<basePluginOptions> = (opts?: basePluginOptions) 
 							return;
 						}
 
+						if (colorData.syntaxFlags.has(SyntaxFlag.Experimental)) {
+							return;
+						}
+
 						if (colorData.syntaxFlags.has(SyntaxFlag.HasNoneKeywords)) {
 							return;
 						}